requirements.txt
google-generativeai>=0.3.0
PyMuPDF>=1.23.0

